Add files via upload
Project on Stock Price Prediction

Time series case studies are one of the most challenging problems in finncial engineering forecast since it adds complexity of a sequence dependence among the input variables. Time series problems could be solved using traditional methods as well as deep learning algorithms. Time series problems are peculiar problems unlike regression analysis and stock price prediction is one such example where deep learning model would do benefit. The project makes use of LSTM model so as to give fair results for opening prices of stocks. For this, initially, RNN and LSTM, both the models would be run and as per the evaluations, LSTM model would be used furtheron to check the deviations of real test data with the predicted test values using the LSTM model. Furthermore, it would also be complimented with proper visualization of plots, and also how much the test data is deviated form original train data set. So for this project, the train set of around 1250 rows would be used and a test set of about 20 rows would be used for checking the error and deviations